When timing out and doing the secondary, don't do the primary action
Only one of the actions should be performed: if one of them triggered, the other
should not. Fortunately for us, due to how the `timer` - `timeOut` comparisons
were set up, all we need to do is let the timer grow a little bigger than the
timeout, and we achieve the desired result.

Fixes #3.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>