Add python based builder script
This script is able to find, list and build the various sketches
in the repo, as well as build all of them.

I also have code for flashing projects, but to test it I'd need
to port in my keyboard firmware and I'd like to reach a resolution
on https://github.com/keyboardio/Kaleidoscope/issues/137 first.